Extract mock to its own file. Move to test target
