+++ WooCommerce/Classes/ViewRelated/Products/ProductsListViewModel.swift
@@ -2,27 +2,10 @@ import Foundation
 import Yosemite
 import Experiments
 
-final class MockProductListViewModel: ProductsListViewModelProtocol {
-    private(set) var scanToUpdateInventoryShouldBeVisible: Bool = false
-    private let featureFlagService: FeatureFlagService
-
-    init(featureFlagService: FeatureFlagService) {
-        self.featureFlagService = featureFlagService
-    }
-
-    func scanToUpdateInventoryButtonShouldBeVisible(completion: @escaping (Bool) -> (Void)) {
-        guard self.featureFlagService.isFeatureFlagEnabled(.scanToUpdateInventory) else {
-            scanToUpdateInventoryShouldBeVisible = false
-            return completion(false)
-        }
-        scanToUpdateInventoryShouldBeVisible = true
-        completion(true)
-    }
-}
-
 protocol ProductsListViewModelProtocol {
     func scanToUpdateInventoryButtonShouldBeVisible(completion: @escaping (Bool) -> (Void))
 }
+
 /// View model for `ProductsViewController`. Has stores logic related to Bulk Editing and Woo Subscriptions.
 ///
 final class ProductListViewModel: ProductsListViewModelProtocol {
